fix: Corregir TODAS las funciones auxiliares para que coincidan con el esquema local - Corregir createSingleTable.orders.id de bigint(20) a int(11) - Corregir createSingleTable.order_items.order_id de bigint(20) a int(11) - Corregir createSingleTable.order_status_history.order_id de bigint(20) a int(11) - Esto asegura que TODAS las funciones de creación de tablas respeten el esquema local - Resuelve el error persistente ER_FK_INCOMPATIBLE_COLUMNS

+++ backend/src/config/database.js
@@ -523,7 +523,7 @@ const createSingleTable = async (connection, tableName) => {
       case 'orders':
         await connection.query(`
           CREATE TABLE IF NOT EXISTS orders (
-            id bigint(20) NOT NULL AUTO_INCREMENT,
+            id int(11) NOT NULL AUTO_INCREMENT,
             order_number varchar(50) NOT NULL,
             customer_type enum('registered','guest') NOT NULL,
             user_id bigint(20) DEFAULT NULL,
@@ -563,7 +563,7 @@ const createSingleTable = async (connection, tableName) => {
         await connection.query(`
           CREATE TABLE IF NOT EXISTS order_items (
             id int(11) NOT NULL AUTO_INCREMENT,
-            order_id bigint(20) NOT NULL,
+            order_id int(11) NOT NULL,
             product_id bigint(20) NOT NULL,
             product_name varchar(255) NOT NULL,
             product_price decimal(10,2) NOT NULL,
@@ -740,7 +740,7 @@ const createSingleTable = async (connection, tableName) => {
         await connection.query(`
           CREATE TABLE IF NOT EXISTS order_status_history (
             id int(11) NOT NULL AUTO_INCREMENT,
-            order_id bigint(20) NOT NULL,
+            order_id int(11) NOT NULL,
             previous_status enum('pending','confirmed','preparing','ready','delivered','cancelled') DEFAULT NULL,
             new_status enum('pending','confirmed','preparing','ready','delivered','cancelled') NOT NULL,
             changed_by enum('system','admin','customer') NOT NULL,
